# util/deephash: hash uint{8,16,32,64} explicitly