@dsnet 2021-07-23
Approximate increase of 28 lines of source code and ~1KiB in compile size. This is offset by #2502